requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
pandas==2.1.3
requests==2.31.0
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.145
opencv-python-headless==4.8.1.78
pillow==10.1.0
huggingface_hub==0.19.0
python-dotenv==1.0.0 